Fixed request, added feed and bmodelcanvas packages

+++ packages/users/src/server/server.ts
@@ -2,10 +2,12 @@
 import { Session, NO_SESSION, User, NO_USER, NO_TOKEN } from '../common';
 import { Server } from "@startupway/main/lib/server";
 import { getPool } from '@startupway/database/lib/server';
-import {QueryOptions, Connection} from 'mariadb';
-import {Router} from "express";
+import { QueryOptions, Connection } from 'mariadb';
+import { NextFunction, Router, Request, Response } from "express";
 import { createHash } from 'crypto';
 import { generate } from 'randomstring';
+import { TABLE_USERS } from './tables';
+
 export class UsersServer {
 
 	private static INSTANCE?: UsersServer;
@@ -23,38 +25,33 @@ export class UsersServer {
 		})
 	}
 
-	_passwordGenerator (password: string): string {
+	private passwordGenerator (password: string): string {
 		const hash = createHash('sha256');
 		hash.update(password);
 		return hash.digest('hex');
 	}
 
-	async addUser(user: User): Promise<User> {
+	async addUser(user: User): Promise<User | null> {
 		try {
-			user.password = this._passwordGenerator (user.password);
+			user.password = this.passwordGenerator (user.password);
 
 			const queryOptions:QueryOptions = {
-				sql: "INSERT INTO users columns(firstName,lastName,username,password,email,phone,birthDate,avatarUu,socialMedia,userDetails,role,lastLogin) values(?,?,?,?,?,?,?,?,?,?,?,?)"
+				namedPlaceholders:true,
+				sql: `INSERT INTO ${TABLE_USERS} COLUMNS(firstName,lastName,username,password,email,phone,birthDate,avatarUu,socialMedia,userDetails,role,lastLogin) values(:firstName,:lastName,username,password,email,phone,birthDate,avatarUu,socialMedia,userDetails,role,lastLogin)`
 			}
-			const values = [
-				user.firstName, user.lastName, user.username,
-				user.password, user.email, user.phone,
-				user.birthDate, user.avatarUu, user.socialMedia,
-				user.userDetails, user.role, user.lastLogin
-			];
-			const response = await this.conn.query(queryOptions,values);
+			const response = await this.conn.query(queryOptions,user);
 			if(response) {
 				const newUser = await this.getUserByUsername(user.username);
 				if(newUser)
 					return newUser;
 				else
-					return NO_USER;
+					return null;
 			} else {
-				return NO_USER;
+				return null;
 			}
 		} catch (error) {
 			console.error(error);
-			return NO_USER;
+			return null;
 		}
 	}
 	async getAllUserTeams(): Promise<User[]> {
@@ -76,9 +73,12 @@ export class UsersServer {
 	async deleteUser(user: User): Promise<void> {
 		try {
 			const queryOptions:QueryOptions = {
-				sql: "DELETE FROM users where userId=?"
+				namedPlaceholders:true,
+				sql: "DELETE FROM users where userId=:userId"
 			}
-			const values:number[] = [user.userId];
+			const values = {
+				userId:user.userId
+			};
 			await this.conn.query(queryOptions,values);
 		} catch (error) {
 			// TODO add user back if failed
@@ -87,36 +87,40 @@ export class UsersServer {
 		}
 	}
 
-	async createSession(username: string, password: string): Promise<Session> {
-		password = this._passwordGenerator (password);
+	async createSession(username: string, password: string): Promise<Session | null> {
+		password = this.passwordGenerator (password);
 		try {
 			let queryOptions:QueryOptions = {
-				sql: "SELECT username FROM users WHERE username=?"
+				namedPlaceholders:true,
+				sql: "SELECT username FROM users WHERE username=:username"
 			}
-			const resUsername:{username:string}[] = await this.conn.query(queryOptions,[username]) as {username:string}[];
+			const resUsername:{username:string}[] = await this.conn.query(queryOptions,{username:username}) as {username:string}[];
 			if(resUsername[0]) {
 				queryOptions = {
-					sql: "SELECT * FROM users WHERE username=? AND password=?"
+					namedPlaceholders:true,
+					sql: "SELECT * FROM users WHERE username=:username AND password=:password"
 				}
-				const user:User[] = await this.conn.query(queryOptions,[username, password]) as User[];
+				const user:User[] = await this.conn.query(queryOptions,{username:username,password:password}) as User[];
 				if(user[0]!== undefined && user[0].userId !== 0) {
 					const userId:number = user[0].userId;
 					const token:string =generate({ length: 100 });
 					queryOptions = {
-						sql: "INSERT INTO sessions (userId, token) values(?,?)"
+						namedPlaceholders:true,
+						sql: "INSERT INTO sessions (userId, token) values(:userId,:token)"
 					}
-					const resSession:Session[] =await this.conn.query(queryOptions,[userId, token]) as Session[];
+					const resSession:Session[] =await this.conn.query(queryOptions,{userId:userId, token:token}) as Session[];
 					if(resSession) {
 						queryOptions = {
-							sql: "SELECT * FROM sessions WHERE userId=?"
+							namedPlaceholders:true,
+							sql: "SELECT * FROM sessions WHERE userId=:userId"
 						}
-						const session:Session[] = await this.conn.query(queryOptions,[userId]) as Session[];
+						const session:Session[] = await this.conn.query(queryOptions,{userId:userId}) as Session[];
 						if(session[0])
 							return session[0];
 						else
-							return NO_SESSION;
+							return null;
 					} else {
-						return NO_SESSION;
+						return null;
 					}
 				} else {
 					return {sessionId:0,token:"cred",userId:0,createdAt: new Date(0)};
@@ -132,100 +136,112 @@ export class UsersServer {
 		}
 	}
 
-	async modifyUser(user: User) {
+	async modifyUser(user: User, changedPass?: string) {
 		try {
-			const queryOptions:QueryOptions = {
-				sql: "UPDATE users SET firstName=?, lastName=?, username=?, password=?,email=?,phone=?,socialMedia=?,birthDate=?,userDetails=?,role=?,avatarUu=?,lastLogin=? where userId=?"
+			if(changedPass) {
+				user.password = this.passwordGenerator(user.password);
 			}
-			const values:(string|number|Date|{[key:string]:string})[] = [
-				user.firstName,user.lastName,user.username,
-				user.password,user.email,user.phone,
-				user.socialMedia,user.birthDate,user.userDetails,
-				user.role,user.avatarUu,user.lastLogin,
-				user.userId
-			];
-			await this.conn.query(queryOptions,values);
+			const queryOptions:QueryOptions = {
+				namedPlaceholders:true,
+				sql: "UPDATE users SET firstName=:firstName, lastName=:lastName, username=:username, password=:password, email=:email, phone=:phone, socialMedia=:socialMedia, birthDate=:birthDate, userDetails=:userDetails, role=:role, avatarUu=:avatarUu, lastLogin=:lastLogin WHERE userId=:userId"
+			}
+			await this.conn.query(queryOptions,user);
 		} catch (error) {
 			console.error(error);
 		}
 	}
 
-	async getUserByUsername(username: string): Promise<User> {
+	async getUserByUsername(username: string): Promise<User | null> {
 		try {
 			const queryOptions:QueryOptions = {
+				namedPlaceholders:true,
 				sql: "SELECT * FROM users WHERE username=?"
 			}
-			const values:string[] = [username];
+			const values = {
+				username:username
+			};
 			const user:User[] = await this.conn.query(queryOptions,values) as User[];
 			if(user.length > 0) {
 				if (user[0].userId !== 0)
 					return user[0];
 				else
-					return NO_USER;
+					return null;
 			} else {
-				return NO_USER;
+				return null;
 			}
 		} catch (error) {
 			console.error(error);
-			return NO_USER;
+			return null;
 		}
 	}
-	async getUserByEmail(email: string): Promise<User> {
+	async getUserByEmail(email: string): Promise<User | null> {
 		try {
 			const queryOptions:QueryOptions = {
+				namedPlaceholders:true,
 				sql: "SELECT * FROM users WHERE email=?"
 			}
-			const values:string[] = [email];
+			const values= {
+				email:email
+			};
 			const user:User[] = await this.conn.query(queryOptions,values) as User[];
 			if(user.length > 0) {
 				if (user[0].userId !== 0)
 					return user[0];
 				else
-					return NO_USER;
+					return null;
 			} else {
-				return NO_USER;
+				return null;
 			}
 		}
 		catch (error) {
 			console.error(error);
-			return NO_USER;
+			return null;
 		}
 	}
 
-	async getUserById(userId: number): Promise<User> {
+	async getUserById(userId: number): Promise<User | null> {
 		try {
 			const queryOptions:QueryOptions = {
+				namedPlaceholders:true,
 				sql: "SELECT * FROM users WHERE userId=?"
 			}
-			const values:number[] = [userId];
+			const values = {
+				userId:userId
+			}
 			const user:User[] = await this.conn.query(queryOptions,values) as User[];
 			if(user.length > 0) {
 				if (user[0].userId !== 0)
 					return user[0];
 				else
-					return NO_USER;
+					return null;
 			} else {
-				return NO_USER;
+				return null;
 			}
 		} catch (error) {
 			console.error(error);
-			return NO_USER;
+			return null;
 		}
 	}
 
 	async deleteSession(token: string, sessionId?: number): Promise<void> {
 		try {
 			const queryOptions:QueryOptions = {
+				namedPlaceholders:true,
 				sql: ""
 			}
-			const values:(string|number)[] = [];
+			let values:{token?:string,sessionId?:number} = {};
 			if(sessionId){
-				queryOptions.sql = "DELETE FROM users WHERE token=? AND sessionID=?";
-				values.push(token,sessionId);
+				queryOptions.sql = "DELETE FROM users WHERE token=:token AND sessionID=:sessionID";
+				values = {
+					token:token,
+					sessionId:sessionId
+				}
 			}
 			else {
-				queryOptions.sql = "DELETE FROM users WHERE token=?";
-				values.push(token);
+				queryOptions.sql = "DELETE FROM users WHERE token=:token";
+				values = {
+					token:token,
+				}
 			}
 			await this.conn.query(queryOptions,values);
 		} catch (error) {
@@ -234,19 +250,22 @@ export class UsersServer {
 	}
 
 
-	async getUserLastSession(userId: number): Promise<Session> {
+	async getUserLastSession(userId: number): Promise<Session | null> {
 		try {
 			const queryOptions:QueryOptions = {
-				sql: "SELECT * FROM session where userId=?"
+				namedPlaceholders:true,
+				sql: "SELECT * FROM session where userId=:userId"
 			}
-			const values:(number)[] = [userId];
+			const values = {
+				userId:userId
+			}
 			const session:Session[] = await this.conn.query(queryOptions,values) as Session[];
 			if(session[0])
 				return session[0];
-			return NO_SESSION;
+			return null;
 		} catch (error) {
 			console.error(error);
-			return NO_SESSION;
+			return null;
 		}
 	}
 
@@ -285,25 +304,26 @@ export class UsersServer {
 		}
 	}
 
-	async getSessionUser(token: string): Promise<User> {
+	async getSessionUser(token: string): Promise<User | null> {
 		try {
 			let queryOptions:QueryOptions = {
-				sql: "SELECT userId FROM sessions where token=?"
+				namedPlaceholders:true,
+				sql: "SELECT userId FROM sessions where token=:token"
 			}
-			const session:{userId:number}[] =  await this.conn.query(queryOptions,[token]) as {userId:number}[];
+			const session:{userId:number}[] =  await this.conn.query(queryOptions,{token:token}) as {userId:number}[];
 			if(session[0]) {
 				queryOptions = {
-					sql: "SELECT * FROM users WHERE userId=?"
+					namedPlaceholders:true,
+					sql: "SELECT * FROM users WHERE userId=:userId"
 				}
-				const user:User[] = await this.conn.query(queryOptions,[session[0].userId]) as User[];// where data de azi mai noua decat expirare
+				const user:User[] = await this.conn.query(queryOptions,{userId:session[0].userId}) as User[];// where data de azi mai noua decat expirare
 				if(user[0])
 					return user[0];
 			}
-			return NO_USER;
+			return null;
 		} catch (error) {
 			console.error(error);
-			return NO_USER;
-
+			return null;
 		}
 	}
 
@@ -322,6 +342,7 @@ export class UsersServer {
 const server = Server.getInstance ();
 const usersServer = UsersServer.getInstance();
 const router = Router ();
+
 router.use((req, res, next) => {
 	(req as any).user = NO_USER;
 	(req as any).token = NO_TOKEN;
@@ -330,13 +351,15 @@ router.use((req, res, next) => {
 
 router.post("/login", async (req, res) => {
 	try {
-		const session:Session = await usersServer.createSession(req.body.username, req.body.password);
+		const session:Session | null = await usersServer.createSession(req.body.username, req.body.password);
 		const user = await usersServer.getUserByUsername(req.body.username);
-		if(user.userId && user.userId !== 0) {
+		if(user) {
 			user.lastLogin = req.body.lastLogin
 			await usersServer.modifyUser(user)
 		}
-		if(session.token === "error") {
+		if(session === null) {
+			res.send("");
+		} else if(session.token === "error") {
 			res.status(500).send({err:500});
 		} else {
 			res.send (session);
@@ -352,7 +375,7 @@ router.get("/verify/:email", async(req,res) => {
 	try {
 		const email = req.params.email;
 		const user = await usersServer.getUserByEmail(email);
-		if(user.userId !== 0) {
+		if(user) {
 			res.status(200).send({accept:"Yes"});
 		} else {
 			res.status(404).send({err:404});
@@ -364,28 +387,10 @@ router.get("/verify/:email", async(req,res) => {
 	}
 });
 
-router.use(async (req, res, next)=> {
-	const authorization = req.header("Authorization");
-	let token = NO_TOKEN;
-	if(authorization){
-		token = authorization.split(" ")[1];
-	}
-	if(token !==  NO_TOKEN) {
-		const user = await usersServer.getSessionUser(token);
-		if(user === NO_USER)
-			res.status(401).send({err:401});
-		else
-		{
-			(req as any).user = user;
-			(req as any).token = token;
-			next();
-		}
-	}
-	else
-	{
-		res.status(401).send({err:401});
-	}
-});
+const authFunct = getAuthorizationFunction();
+if(authFunct)
+	router.use((authFunct as any));
+	// Bypass params dictionary and send authorization Function
 
 router.get("/user", async (req, res, next) => {
 	res.send((req as any).user);
@@ -393,11 +398,10 @@ router.get("/user", async (req, res, next) => {
 
 
 router.post("/user/update", async (req,res) => {
-	const user = req.body.newUser;
+	const user:User = req.body.newUser;
 	const changedPass = req.body.changedPass;
-	if(changedPass) {
-		user.password = usersServer._passwordGenerator(user.password);
-	}
+	console.log("Changed" + changedPass);
+	console.log(user);
 	// const options = admin.createMailOptions(
 	// 	(process.env.MAIL_USER as string),
 	// 	user.email,
@@ -413,7 +417,7 @@ router.post("/user/update", async (req,res) => {
 	// admin.sendMail(transporter,options);
 
 	if(user) {
-		await usersServer.modifyUser((user as User));
+		await usersServer.modifyUser(user, changedPass);
 	} else {
 		res.status(401).send({err:401});
 	}
@@ -456,7 +460,7 @@ router.get("/users/all", async (req,res) => {
 	}
 });
 
-router.post("/users/logout", async (req, res, next) => {
+router.post("/logout", async (req, res, next) => {
 	if(req.body.sessionId)
 		await usersServer.deleteSession((req as any).token, req.body.sessionId);
 	else
@@ -466,7 +470,7 @@ router.post("/users/logout", async (req, res, next) => {
 router.get("/session/:userId", async (req,res) => {
 	try {
 		// const userId = req.params.userId;
-		const session: Session = await usersServer.getUserLastSession(Number(req.params.userId));
+		const session: Session | null = await usersServer.getUserLastSession(Number(req.params.userId));
 		if(session) {
 			res.status(200).send(session);
 		} else {
@@ -478,3 +482,35 @@ router.get("/session/:userId", async (req,res) => {
 	}
 });
 server.registerRouterAPI (1, router, "/users");
+
+
+export function getAuthorizationFunction(): Function | null {
+	try {
+		let f = async (req:Request, res:Response, next:NextFunction)=> {
+			const authorization = req.header("Authorization");
+			let token = NO_TOKEN;
+			if(authorization){
+				token = authorization.split(" ")[1];
+			}
+			if(token !==  NO_TOKEN) {
+				const user = await usersServer.getSessionUser(token);
+				if(user === NO_USER)
+					res.status(401).send({err:401});
+				else
+				{
+					(req as any).user = user;
+					(req as any).token = token;
+					next();
+				}
+			}
+			else
+			{
+				res.status(401).send({err:401});
+			}
+		}
+		return f;
+	} catch(error) {
+		console.error(error);
+		return null;
+	}
+};
